test(statistics): split AnalysisExplorerShellTest by UI behaviour
Separate happy-path, filter, auth/metadata, and edge-case LiveComponent coverage; mark the oversized-class backlog done.
